commit favorites functionality worked on trevor/tarik/JD tadlock

+++ public/js/favorites.js
@@ -2,26 +2,40 @@
 const deleteButton = document.querySelector('#favorite-books')
 
 
+
 //event listener when clicked on delete
-deleteButton.addEventListener('click', async (e) => {
-  if (e.target.classList.contains('delete-button')) {
-    const bookElementId = e.target.closest('.book').dataset.id  //  datasetId 
-    const deleteHtml = e.target.closest('.book')
-    
-    // DELETE HTML
-    deleteHtml.remove()
+if (deleteButton) {
+    deleteButton.addEventListener('click', async (e) => {
+        if (e.target.classList.contains('delete-button')) {
+            const bookElementId = e.target.closest('.book').dataset.id  //  datasetId 
+            const deleteHtml = e.target.closest('.book')
 
-    //target the data-id="{{id}}" and call(fetch) the DELETE route 
-    const bookData = {
-      api_id: bookElementId
-    }
+            // DELETE HTML
+            deleteHtml.remove()
 
-    await fetch('/favorites', {
-      method: 'DELETE',
-      headers: {
-        'Content-Type': 'application/json'
-      },
-      body: JSON.stringify(bookData)  //api_id
+            //target the data-id="{{id}}" and call(fetch) the DELETE route 
+            const bookData = {
+                api_id: bookElementId
+            }
+
+            try {
+              
+                await fetch('/favorites', {
+                    method: 'DELETE',
+                    headers: {
+                        'Content-Type': 'application/json'
+                    },
+                    body: JSON.stringify(bookData)  //api_id
+                })
+
+                if (!deleteButton.children.length) {
+                    window.location.reload()
+                }
+            } catch (error) {
+                console.log(error)
+            }
+
+        }
     })
-  }
-})
+}
+
